docs/agentic/execution/requirements.txt
# Python dependencies for execution scripts
# Install with: pip install -r requirements.txt

requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0
